@@ -20,7 +20,6 @@ package database
 import (
 	"encoding/json"
 	"errors"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"strconv"
@@ -71,7 +70,7 @@ func init() {
 func TestMain(m *testing.M) {
 	var err error
 
-	tmpDir, err = ioutil.TempDir("", "sm_")
+	tmpDir, err = os.MkdirTemp("", "sm_")
 	if err != nil {
 		log.Fatalf("Error create temporary dir: %s", err)
 	}
