@@ -23,7 +23,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -107,7 +106,7 @@ func TestFileServer(t *testing.T) {
 
 	filename := "testFile.txt"
 
-	if err := ioutil.WriteFile(filepath.Join(serverDir, filename), []byte("Hello fileserver"), 0o600); err != nil {
+	if err := os.WriteFile(filepath.Join(serverDir, filename), []byte("Hello fileserver"), 0o600); err != nil {
 		t.Fatalf("Can't create package file: %s", err)
 	}
 
@@ -123,7 +122,7 @@ func TestFileServer(t *testing.T) {
 	time.Sleep(1 * time.Second)
 
 	// Get the data
-	resp, err := http.Get(outURL) // nolint // just for tests
+	resp, err := http.Get(outURL) //nolint:gosec
 	if err != nil {
 		t.Fatalf("Can't download file: %s", err)
 	}
