@@ -24,7 +24,6 @@ import (
 	"errors"
 	"fmt"
 	"html/template"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -71,13 +70,19 @@ type dnsServer struct {
  **********************************************************************************************************************/
 
 // These global variable is used to be able to mocking the functionality of networking in tests.
-// nolint:gochecknoglobals
+//
+//nolint:gochecknoglobals
 var (
 	LookPath          = exec.LookPath
 	DiscoverInterface = gateway.DiscoverInterface
 	ExecContext       = execShellCommander
-	errProcessNotExit = aoserrors.New("process is not exit")
 )
+
+var errProcessNotExist = aoserrors.New("process not exist")
+
+/***********************************************************************************************************************
+ * Private
+ **********************************************************************************************************************/
 
 func newDNSServer(networkDir string) (*dnsServer, error) {
 	dnsMasqBinary, err := LookPath("dnsmasq")
@@ -177,7 +182,7 @@ func (dns *dnsServer) prepareDNSConfFile() error {
 		return aoserrors.Wrap(err)
 	}
 
-	return aoserrors.Wrap(ioutil.WriteFile(dns.configFile, newConfig, 0o600))
+	return aoserrors.Wrap(os.WriteFile(dns.configFile, newConfig, 0o600))
 }
 
 func (dns *dnsServer) generateDNSMasqConfig() ([]byte, error) {
@@ -199,7 +204,7 @@ func (dns *dnsServer) generateDNSMasqConfig() ([]byte, error) {
 
 func (dns *dnsServer) restart() error {
 	process, err := dns.findServerProcess()
-	if err != nil && !errors.Is(err, errProcessNotExit) {
+	if err != nil && !errors.Is(err, errProcessNotExist) {
 		return aoserrors.Wrap(err)
 	}
 
@@ -213,15 +218,19 @@ func (dns *dnsServer) restart() error {
 func (dns *dnsServer) findServerProcess() (*os.Process, error) {
 	if _, err := os.Stat(dns.PidFile); err != nil {
 		if os.IsNotExist(err) {
-			return nil, errProcessNotExit
+			return nil, errProcessNotExist
 		}
 
 		return nil, aoserrors.Wrap(err)
 	}
 
-	pidFileContents, err := ioutil.ReadFile(dns.PidFile)
+	pidFileContents, err := os.ReadFile(dns.PidFile)
 	if err != nil {
 		return nil, aoserrors.Wrap(err)
+	}
+
+	if len(pidFileContents) == 0 {
+		return nil, errProcessNotExist
 	}
 
 	pid, err := strconv.Atoi(strings.TrimSpace(string(pidFileContents)))
