@@ -57,7 +57,8 @@ type MonitorController struct {
  **********************************************************************************************************************/
 
 // MinSendPeriod used to adjust min send period.
-// nolint:gochecknoglobals
+//
+//nolint:gochecknoglobals
 var MinSendPeriod = 10 * time.Second
 
 /***********************************************************************************************************************
